refactor: move position_in_list to manager.py
