+++ src/CreeDictionary/CreeDictionary/paradigm/sort_utils.py
@@ -36,13 +36,3 @@ SupportsLessThan = TypeVar("SupportsLessThan", bound=_SupportsLessThan)
 KeyFunction = Callable[[_T], SupportsLessThan]
 
 
-def position_in_list(reference: list[str]) -> Callable[[str], int]:
-    """
-    Returns a key function that will sort an element by its position in the given list.
-    """
-    map_element_to_index = {element: index for index, element in enumerate(reference)}
-
-    def key_function(element: str):
-        return map_element_to_index[element]
-
-    return key_function
